+++ back/src/jacket-client/models/jcategory.model.ts
@@ -1,0 +1,5 @@
+export enum JCategory {
+  MOVIES = 2000,
+  AUDIO = 3000,
+  SERIES = 5000,
+}
