+++ back/src/jacket-client/models/jsearch.model.ts
@@ -1,0 +1,5 @@
+export interface JSearch {
+  query: string;
+  categories?: any[];
+  tracker?: string;
+}
